requirements.txt
psycopg2-binary
python-dotenv
requests
openpyxl
beautifulsoup4
geopy
